refactor(core): :recycle: promote self-signed cert minter to core lib/tls-cert.mjs
